Give user back control when events are exhuasted
Instead of behaving like a Stream from the futures crate, EventStream
should be 'restartable.' Every frame the user can exhaust all the
available events and receive 'None', process their updates and re-draw
based on the events, and then wait for more events.

The Stream API views None as a permanent termination point, which is
incompatible with this paradigm. This removes the usage of Stream,
replaced instead with a function that generates a Future directly.